Refactor: 주문 관련 컴포저블 구조 분리 및 UI 개선
주문하기 관련 컴포저블(`OrderButton`, `GotoOrderRow`)을 `order` 패키지로 분리하여 구조를 개선하고, 장바구니 화면의 전반적인 UI 간격과 패딩을 조정했습니다.

- **`OrderButton.kt`**
    - `GotoOrderRow.kt`에 있던 `OrderButton` 컴포저블을 `order` 패키지 내의 새로운 파일로 분리하여 재사용성을 높였습니다.

- **`GotoOrderRow.kt`**
    - 파일을 `component/`에서 `component/order/` 디렉토리로 이동시켰습니다.
    - `OrderButton`이 별도 파일로 분리됨에 따라, 관련 코드를 삭제했습니다.

- **`CartScreen.kt`**
    - 파일 경로 변경에 따라 `GotoOrderRow`의 import 경로를 수정했습니다.
    - '결제금액' 섹션의 제목 위아래와 카드 하단에 `Spacer`를 추가하여 UI 간격을 조정했습니다.

- **`TotalAmountCard.kt`**
    - 결제 금액 카드의 내부 패딩을 `16.dp`에서 `20.dp`로 수정하고, 불필요한 `Spacer`들을 제거하여 레이아웃을 개선했습니다.
